functions revised final

+++ CartaMetrica/core/function.js
@@ -29,8 +29,8 @@ export function getCoords(lines){
     
     // calculate latitude for each line
     for (let i = 0; i < lines.length; i++) {
-        let distance = parseInt(lines[i].distance );
-        let azimuth_n = parseInt(lines[i].azimuth );
+        let distance =  parseFloat(lines[i].distance );
+        let azimuth_n =  parseFloat(lines[i].azimuth );
         let latitude = (distance*cos(radians*azimuth_n))
         lat_of_lines.push(latitude);
     }
@@ -58,8 +58,8 @@ export function getCoords(lines){
     
     // calculate departure for each line
     for (let i1 = 0; i1 < lines.length; i1++) {
-        let distance = parseInt(lines[i1].distance );
-        let azimuth_n = parseInt(lines[i1].azimuth );
+        let distance =  parseFloat(lines[i1].distance );
+        let azimuth_n = parseFloat(lines[i1].azimuth );
         let departure = (distance*sin(radians*azimuth_n))
         dep_of_lines.push(departure);
     }
